\name{hnp-internal}
\alias{.makehnp}
\encoding{UTF-8}
\title{Internal function to prepare \code{hnp} objects}
\description{This is an internally called function used to prepare \code{hnp} objects for plotting}
\author{
Rafael A. Moral <rafael_moral@yahoo.com.br>, John Hinde and Clarice G. B. \enc{Demétrio}{Demetrio}
}